style: adjust layout
