+++ src/app/[lang]/series/page.tsx
@@ -1,14 +1,10 @@
 import { getTitle } from '~/helpers/get-title';
 import { Metadata } from 'next';
-import { translate } from '~/helpers/i18n';
 import { Language, SUPPORTED_LANGUAGES } from '~/helpers/i18n/config';
 import { notFound } from 'next/navigation';
 import { getAllSeries } from '~/helpers/get-series';
 import SeriesList from '~/components/Series/List';
-import en from '../../../../public/locales/en/common.json';
-import zh from '../../../../public/locales/zh/common.json';
-
-const translations: Record<Language, typeof en> = { en, zh };
+import { createTranslator } from '~/helpers/i18n/translations';
 
 interface SeriesPageProps {
   params: Promise<{
@@ -18,7 +14,7 @@ interface SeriesPageProps {
 
 export async function generateMetadata({ params }: SeriesPageProps): Promise<Metadata> {
   const { lang: paramLang } = await params;
-  const t = (key: string) => translate(key, paramLang as Language, translations);
+  const t = createTranslator(paramLang as Language);
   const title = getTitle(t('meta_series_list_title'));
   
   return {
@@ -47,7 +43,7 @@ export default async function SeriesPage({ params }: SeriesPageProps) {
   }
 
   const lang = paramLang as Language;
-  const t = (key: string) => translate(key, lang, translations);
+  const t = createTranslator(lang);
   const series = getAllSeries(lang);
 
   return (
